+++ crates/swc_ecma_ast/src/decl.rs
@@ -175,8 +175,13 @@ impl Take for VarDeclarator {
 #[derive(Eq, Hash, EqIgnoreSpan)]
 #[cfg_attr(feature = "arbitrary", derive(arbitrary::Arbitrary))]
 pub struct UsingDecl {
+    #[cfg_attr(feature = "serde-impl", serde(default))]
     pub span: Span,
 
+    #[cfg_attr(feature = "serde-impl", serde(default))]
+    pub is_await: bool,
+
+    #[cfg_attr(feature = "serde-impl", serde(default))]
     pub decls: Vec<VarDeclarator>,
 }
 
@@ -184,6 +189,7 @@ impl Take for UsingDecl {
     fn dummy() -> Self {
         Self {
             span: DUMMY_SP,
+            is_await: Default::default(),
             decls: Take::dummy(),
         }
     }