+++ crates/swc_ecma_parser/src/parser/stmt.rs
@@ -103,6 +103,7 @@ impl<'a, I: Tokens> Parser<I> {
 
         let ctx = Context {
             will_expect_colon_for_cond: false,
+            allow_using_decl: true,
             ..self.ctx()
         };
         self.with_ctx(ctx)
@@ -132,6 +133,15 @@ impl<'a, I: Tokens> Parser<I> {
                 self.emit_err(self.input.cur_span(), SyntaxError::TopLevelAwaitInScript);
             }
 
+            if peeked_is!(self, "using") {
+                assert_and_bump!(self, "await");
+
+                let v = self.parse_using_decl(start, true)?;
+                if let Some(v) = v {
+                    return Ok(Stmt::Decl(Decl::Using(v)));
+                }
+            }
+
             let expr = self.parse_await_expr()?;
             let expr = self
                 .include_in_expr(true)
@@ -154,6 +164,16 @@ impl<'a, I: Tokens> Parser<I> {
         }
 
         match cur!(self, true)? {
+            tok!("await") if include_decl => {
+                if peeked_is!(self, "using") {
+                    assert_and_bump!(self, "await");
+                    let v = self.parse_using_decl(start, true)?;
+                    if let Some(v) = v {
+                        return Ok(Stmt::Decl(Decl::Using(v)));
+                    }
+                }
+            }
+
             tok!("break") | tok!("continue") => {
                 let is_break = is!(self, "break");
                 bump!(self);
@@ -302,7 +322,7 @@ impl<'a, I: Tokens> Parser<I> {
             }
 
             tok!("using") if include_decl => {
-                let v = self.parse_using_decl()?;
+                let v = self.parse_using_decl(start, false)?;
                 if let Some(v) = v {
                     return Ok(Stmt::Decl(Decl::Using(v)));
                 }
@@ -762,7 +782,11 @@ impl<'a, I: Tokens> Parser<I> {
         }
     }
 
-    pub(super) fn parse_using_decl(&mut self) -> PResult<Option<Box<UsingDecl>>> {
+    pub(super) fn parse_using_decl(
+        &mut self,
+        start: BytePos,
+        is_await: bool,
+    ) -> PResult<Option<Box<UsingDecl>>> {
         // using
         // reader = init()
 
@@ -776,7 +800,6 @@ impl<'a, I: Tokens> Parser<I> {
             return Ok(None);
         }
 
-        let start = cur_pos!(self);
         assert_and_bump!(self, "using");
 
         let mut decls = vec![];
@@ -822,6 +845,7 @@ impl<'a, I: Tokens> Parser<I> {
 
         Ok(Some(Box::new(UsingDecl {
             span: span!(self, start),
+            is_await,
             decls,
         })))
     }
@@ -1278,6 +1302,7 @@ impl<'a, I: Tokens> Parser<I> {
 
             let pat = Box::new(UsingDecl {
                 span: span!(self, start),
+                is_await: false,
                 decls: vec![decl],
             });