+++ crates/swc_ecma_transforms_proposal/tests/explicit-resource-management/transform-sync/for-await-head/output.mjs
@@ -1,0 +1,13 @@
+for await (const x of y){
+    try {
+        var _stack = [];
+        {
+            doSomethingWith(x);
+        }
+    } catch (_) {
+        var _error = _;
+        var _hasError = true;
+    } finally{
+        _dispose(_stack, _error, _hasError);
+    }
+}