+++ crates/swc_ecma_transforms_proposal/tests/explicit-resource-management/transform-top-level/await-or-not-preserved/output.mjs
@@ -1,0 +1,11 @@
+export { x, y };
+try {
+    var _stack = [];
+    var x = _using(_stack, A);
+    var y = _using(_stack, B, true);
+} catch (_) {
+    var _error = _;
+    var _hasError = true;
+} finally{
+    await _dispose(_stack, _error, _hasError);
+}